moved routes to a separate file
