+++ backend/index.js
@@ -2,6 +2,7 @@ import express from "express";
 import { PORT, mongoDBURL } from "./config.js";
 import mongoose from "mongoose";
 import { Book } from "./models/bookModel.js";
+import booksRoute from "./routes/booksRoute.js"
 
 const app = express();
 
@@ -15,118 +16,10 @@ app.get("/", (request, response) => {
 });
 
 
+// Routes
+app.use("/books", booksRoute);
 
-
-
-// Route for Saving a new Book
-app.post("/books", async (request, response) => {
-  try {
-    if (
-      !request.body.title ||
-      !request.body.author ||
-      !request.body.publishYear
-    ) {
-      return response.status(400).send({
-        message: "All fields are required",
-      });
-    }
-    const newBook = {
-      title: request.body.title,
-      author: request.body.author,
-      publishYear: request.body.publishYear,
-    };
-    const book = await Book.create(newBook);
-    return response.status(201).send(book);
-  } catch (error) {
-    return response.status(500).send({ message: error.message });
-  }
-});
-
-
-
-
-// Route for getting all books from MongoDB
-app.get("/books", async (request, response) => {
-  try {
-    const books = await Book.find({});
-    return response.status(200).send({
-      count: books.length,
-      data: books,
-    });
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-
-
-
-// Route to get an individual book by ID from MongoDB
-app.get("/books/:id", async (request, response) => {
-  try {
-    const { id } = request.params;
-
-    const book = await Book.findById(id);
-
-    return response.status(200).send(book);
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-
-
-
-
-// Route to Update a Book
-app.put("/books/:id", async (request, response) => {
-  try {
-    const { id } = request.params;
-    const updates = request.body;
-
-    // Validate if any update fields are provided
-    if (Object.keys(updates).length === 0) {
-      return response.status(400).send({ message: "No update data provided" });
-    }
-
-    const updatedBook = await Book.findByIdAndUpdate(id, updates, {
-      new: true,
-    });
-
-    if (!updatedBook) {
-      return response.status(404).send({ message: "Book not found" });
-    }
-
-    return response.status(200).send(updatedBook);
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-
-
-// Route to Delete a Book
-app.delete("/books/:id", async(request, response) => {
-  try {
-    const { id } = request.params;
-    const result = await Book.findByIdAndDelete(id);
-
-    if(!result){
-      return response.status(404).send({message: "Book not found"});
-    }
-    return response.status(200).send({message: "Book deleted successfully"});
-  } catch(error) {
-    console.log(error.message);
-    response.status(500).send({message: error.message});
-  }
-})
-
-
-
-
+// Connect to MongoDB
 mongoose
   .connect(mongoDBURL)
   .then(() => {
